# Revert "Add histograms for optimizer cost calculation"