--- a/presto-iceberg/src/main/java/com/facebook/presto/iceberg/IcebergSessionProperties.java
+++ b/presto-iceberg/src/main/java/com/facebook/presto/iceberg/IcebergSessionProperties.java
@@ -18,10 +18,9 @@ import com.facebook.presto.hive.HiveCompressionCodec;
 import com.facebook.presto.hive.OrcFileWriterConfig;
 import com.facebook.presto.hive.ParquetFileWriterConfig;
 import com.facebook.presto.iceberg.nessie.NessieConfig;
-import com.facebook.presto.iceberg.util.StatisticsUtil;
+import com.facebook.presto.iceberg.util.HiveStatisticsMergeStrategy;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.session.PropertyMetadata;
-import com.facebook.presto.spi.statistics.ColumnStatisticType;
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import io.airlift.units.DataSize;
@@ -29,13 +28,10 @@ import org.apache.parquet.column.ParquetProperties;
 
 import javax.inject.Inject;
 
-import java.util.EnumSet;
 import java.util.List;
 
 import static com.facebook.presto.common.type.VarcharType.VARCHAR;
 import static com.facebook.presto.common.type.VarcharType.createUnboundedVarcharType;
-import static com.facebook.presto.iceberg.util.StatisticsUtil.SUPPORTED_MERGE_FLAGS;
-import static com.facebook.presto.iceberg.util.StatisticsUtil.decodeMergeFlags;
 import static com.facebook.presto.spi.session.PropertyMetadata.booleanProperty;
 import static com.facebook.presto.spi.session.PropertyMetadata.doubleProperty;
 import static com.facebook.presto.spi.session.PropertyMetadata.integerProperty;
@@ -160,14 +156,14 @@ public final class IcebergSessionProperties
                         false),
                 new PropertyMetadata<>(
                         HIVE_METASTORE_STATISTICS_MERGE_STRATEGY,
-                        "Flags to choose which statistics from the Hive Metastore are used when calculating table stats. Valid values are: "
-                                + Joiner.on(", ").join(SUPPORTED_MERGE_FLAGS),
+                        "choose how to include statistics from the Hive Metastore when calculating table stats. Valid values are: "
+                                + Joiner.on(", ").join(HiveStatisticsMergeStrategy.values()),
                         VARCHAR,
-                        EnumSet.class,
-                        icebergConfig.getHiveStatisticsMergeFlags(),
+                        HiveStatisticsMergeStrategy.class,
+                        icebergConfig.getHiveStatisticsMergeStrategy(),
                         false,
-                        val -> decodeMergeFlags((String) val),
-                        StatisticsUtil::encodeMergeFlags),
+                        val -> HiveStatisticsMergeStrategy.valueOf((String) val),
+                        HiveStatisticsMergeStrategy::name),
                 booleanProperty(
                         PUSHDOWN_FILTER_ENABLED,
                         "Experimental: Enable Filter Pushdown for Iceberg. This is only supported with Native Worker.",
@@ -276,9 +272,9 @@ public final class IcebergSessionProperties
         return session.getProperty(MERGE_ON_READ_MODE_ENABLED, Boolean.class);
     }
 
-    public static EnumSet<ColumnStatisticType> getHiveStatisticsMergeStrategy(ConnectorSession session)
+    public static HiveStatisticsMergeStrategy getHiveStatisticsMergeStrategy(ConnectorSession session)
     {
-        return session.getProperty(HIVE_METASTORE_STATISTICS_MERGE_STRATEGY, EnumSet.class);
+        return session.getProperty(HIVE_METASTORE_STATISTICS_MERGE_STRATEGY, HiveStatisticsMergeStrategy.class);
     }
 
     public static boolean isPushdownFilterEnabled(ConnectorSession session)