--- a/presto-iceberg/src/main/java/com/facebook/presto/iceberg/util/HiveStatisticsMergeStrategy.java
+++ b/presto-iceberg/src/main/java/com/facebook/presto/iceberg/util/HiveStatisticsMergeStrategy.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.iceberg.util;
+
+/**
+ * strategies that define how to merge hive column statistics into Iceberg column statistics.
+ */
+public enum HiveStatisticsMergeStrategy
+{
+    /**
+     * Do not merge statistics from Hive
+     */
+    NONE,
+    /**
+     * Only merge NDV statistics from hive
+     */
+    USE_NDV,
+    /**
+     * Only merge null fractions from hive
+     */
+    USE_NULLS_FRACTIONS,
+    /**
+     * Merge both null fractions and NDVs from Hive
+     */
+    USE_NULLS_FRACTION_AND_NDV,
+}