--- a/presto-main/src/main/java/com/facebook/presto/cost/ConnectorFilterStatsCalculatorService.java
+++ b/presto-main/src/main/java/com/facebook/presto/cost/ConnectorFilterStatsCalculatorService.java
@@ -14,8 +14,6 @@
 
 package com.facebook.presto.cost;
 
-import com.facebook.presto.FullConnectorSession;
-import com.facebook.presto.Session;
 import com.facebook.presto.common.type.Type;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ConnectorSession;
@@ -55,12 +53,7 @@ public class ConnectorFilterStatsCalculatorService
             Map<String, Type> columnTypes)
     {
         PlanNodeStatsEstimate tableStats = toPlanNodeStats(tableStatistics, columnNames, columnTypes);
-        // TODO: Consider re-designing the filter calculator APIs so that a proper Session instance
-        //  can be more easily populated
-        Optional<Session> filterSession = session instanceof FullConnectorSession ?
-                Optional.of(((FullConnectorSession) session).getSession()) :
-                Optional.empty();
-        PlanNodeStatsEstimate filteredStats = filterStatsCalculator.filterStats(tableStats, predicate, session, filterSession);
+        PlanNodeStatsEstimate filteredStats = filterStatsCalculator.filterStats(tableStats, predicate, session);
 
         if (filteredStats.isOutputRowCountUnknown()) {
             filteredStats = tableStats.mapOutputRowCount(sourceRowCount -> tableStats.getOutputRowCount() * UNKNOWN_FILTER_COEFFICIENT);