--- a/presto-main/src/main/java/com/facebook/presto/cost/StatsUtil.java
+++ b/presto-main/src/main/java/com/facebook/presto/cost/StatsUtil.java
@@ -90,7 +90,6 @@ final class StatsUtil
             result.setLowValue(range.getMin());
             result.setHighValue(range.getMax());
         });
-        columnStatistics.getHistogram().ifPresent(result::setHistogram);
         return result.build();
     }
 }