… presto-benchto-benchmarks/src/test/java/com/facebook/presto/sql/planner/AbstractCostBasedPlanTest.java
… presto-benchto-benchmarks/src/test/resources/sql/presto/tpcds/histogram/q85.plan.txt
… presto-docs/src/main/sphinx/connector/iceberg.rst
… presto-hive/src/test/java/com/facebook/presto/hive/TestHiveIntegrationSmokeTest.java
… presto-hive/src/test/java/com/facebook/presto/hive/TestParquetDistributedQueries.java
… presto-hive/src/test/java/com/facebook/presto/hive/TestShowStats.java
… presto-iceberg/src/main/java/com/facebook/presto/iceberg/util/StatisticsUtil.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/IcebergDistributedSmokeTestBase.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/IcebergDistributedTestBase.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/TestIcebergConfig.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/TestIcebergTableChangelog.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/TestStatisticsUtil.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/hive/TestIcebergDistributedHive.java
… presto-iceberg/src/test/java/com/facebook/presto/iceberg/hive/TestIcebergHiveStatistics.java
… presto-main/src/main/java/com/facebook/presto/cost/ComparisonStatsCalculator.java
… presto-main/src/main/java/com/facebook/presto/cost/DisjointRangeDomainHistogram.java
… presto-main/src/main/java/com/facebook/presto/cost/FilterStatsCalculator.java
… presto-main/src/main/java/com/facebook/presto/cost/HistogramCalculator.java
… presto-main/src/main/java/com/facebook/presto/cost/JoinStatsRule.java
… presto-main/src/main/java/com/facebook/presto/cost/PlanNodeStatsEstimateMath.java
… presto-main/src/main/java/com/facebook/presto/cost/StatisticRange.java
… presto-main/src/main/java/com/facebook/presto/cost/UniformDistributionHistogram.java
… presto-main/src/main/java/com/facebook/presto/cost/VariableStatsEstimate.java
… presto-main/src/main/java/com/facebook/presto/sql/rewrite/ShowStatsRewrite.java
… presto-main/src/test/java/com/facebook/presto/cost/TestComparisonStatsCalculator.java
… presto-main/src/test/java/com/facebook/presto/cost/TestComparisonStatsCalculatorHistograms.java
… presto-main/src/test/java/com/facebook/presto/cost/TestComparisonStatsCalculatorNoHistograms.java
… presto-main/src/test/java/com/facebook/presto/cost/TestConnectorFilterStatsCalculatorService.java
… presto-main/src/test/java/com/facebook/presto/cost/TestDisjointRangeDomainHistogram.java
… presto-main/src/test/java/com/facebook/presto/cost/TestFilterStatsCalculator.java