@ZacBlanco 2024-05-03
Thanks for identifying this. I found the root cause. I'll see to fixing it while this is reverted.